+++ src/iter/scheduler/mod.rs
@@ -1,5 +1,10 @@
-use super::plumbing::*;
+//! This module is dedicated to custom scheduler API and useful schedulers.
+
 use super::*;
+
+pub mod misc;
+
+pub use misc::*;
 
 /// Scheduler for Indexed Parallel Iterator
 pub trait Scheduler: Send {
@@ -17,32 +22,6 @@ pub trait UnindexedScheduler: Send {
     where
         P: UnindexedProducer<Item = T>,
         C: UnindexedConsumer<T>;
-}
-
-/// Default Scheduler.
-/// When used as Indexed Scheduler, Thief-splitting will be used.
-/// When used as Unindexed Scheduler, tasks will be divided to minimum piece.
-#[derive(Debug, Clone, Default)]
-pub struct DefaultScheduler;
-
-impl Scheduler for DefaultScheduler {
-    fn bridge<P, C, T>(&mut self, len: usize, producer: P, consumer: C) -> C::Result
-    where
-        P: Producer<Item = T>,
-        C: Consumer<T>,
-    {
-        bridge_producer_consumer(len, producer, consumer)
-    }
-}
-
-impl UnindexedScheduler for DefaultScheduler {
-    fn bridge_unindexed<P, C, T>(&mut self, producer: P, consumer: C) -> C::Result
-    where
-        P: UnindexedProducer<Item = T>,
-        C: UnindexedConsumer<T>,
-    {
-        bridge_unindexed(producer, consumer)
-    }
 }
 
 /// `WithScheduler` is an iterator that enclose one Indexed Scheduler.
